requirements.txt
# ─────────────────────────────────────────────
#  Dashboard Marché du Jeu Vidéo — Dépendances
#  Python 3.9+
# ─────────────────────────────────────────────

# Collecte de données
requests>=2.31.0
requests-cache>=1.1.0
python-dotenv>=1.0.0

# Traitement de données
pandas>=2.1.0
numpy>=1.26.0

# Visualisation
streamlit>=1.35.0
plotly>=5.20.0

# Utilitaires
tqdm>=4.66.0
colorama>=0.4.6

